# Movegen speedup via magic bitboards