--- a/src/chess/bitboard.cc
+++ b/src/chess/bitboard.cc
@@ -269,55 +269,6 @@ const Move kIdxToMove[] = {
     "g7g8b", "g7h8q", "g7h8r", "g7h8b", "h7g8q", "h7g8r", "h7g8b", "h7h8q",
     "h7h8r", "h7h8b"};
 
-// Which squares can rook attack from every of squares.
-static const BitBoard kRookAttacks[] = {
-    0x01010101010101FEULL, 0x02020202020202FDULL, 0x04040404040404FBULL,
-    0x08080808080808F7ULL, 0x10101010101010EFULL, 0x20202020202020DFULL,
-    0x40404040404040BFULL, 0x808080808080807FULL, 0x010101010101FE01ULL,
-    0x020202020202FD02ULL, 0x040404040404FB04ULL, 0x080808080808F708ULL,
-    0x101010101010EF10ULL, 0x202020202020DF20ULL, 0x404040404040BF40ULL,
-    0x8080808080807F80ULL, 0x0101010101FE0101ULL, 0x0202020202FD0202ULL,
-    0x0404040404FB0404ULL, 0x0808080808F70808ULL, 0x1010101010EF1010ULL,
-    0x2020202020DF2020ULL, 0x4040404040BF4040ULL, 0x80808080807F8080ULL,
-    0x01010101FE010101ULL, 0x02020202FD020202ULL, 0x04040404FB040404ULL,
-    0x08080808F7080808ULL, 0x10101010EF101010ULL, 0x20202020DF202020ULL,
-    0x40404040BF404040ULL, 0x808080807F808080ULL, 0x010101FE01010101ULL,
-    0x020202FD02020202ULL, 0x040404FB04040404ULL, 0x080808F708080808ULL,
-    0x101010EF10101010ULL, 0x202020DF20202020ULL, 0x404040BF40404040ULL,
-    0x8080807F80808080ULL, 0x0101FE0101010101ULL, 0x0202FD0202020202ULL,
-    0x0404FB0404040404ULL, 0x0808F70808080808ULL, 0x1010EF1010101010ULL,
-    0x2020DF2020202020ULL, 0x4040BF4040404040ULL, 0x80807F8080808080ULL,
-    0x01FE010101010101ULL, 0x02FD020202020202ULL, 0x04FB040404040404ULL,
-    0x08F7080808080808ULL, 0x10EF101010101010ULL, 0x20DF202020202020ULL,
-    0x40BF404040404040ULL, 0x807F808080808080ULL, 0xFE01010101010101ULL,
-    0xFD02020202020202ULL, 0xFB04040404040404ULL, 0xF708080808080808ULL,
-    0xEF10101010101010ULL, 0xDF20202020202020ULL, 0xBF40404040404040ULL,
-    0x7F80808080808080ULL};
-// Which squares can bishop attack.
-static const BitBoard kBishopAttacks[] = {
-    0x8040201008040200ULL, 0x0080402010080500ULL, 0x0000804020110A00ULL,
-    0x0000008041221400ULL, 0x0000000182442800ULL, 0x0000010204885000ULL,
-    0x000102040810A000ULL, 0x0102040810204000ULL, 0x4020100804020002ULL,
-    0x8040201008050005ULL, 0x00804020110A000AULL, 0x0000804122140014ULL,
-    0x0000018244280028ULL, 0x0001020488500050ULL, 0x0102040810A000A0ULL,
-    0x0204081020400040ULL, 0x2010080402000204ULL, 0x4020100805000508ULL,
-    0x804020110A000A11ULL, 0x0080412214001422ULL, 0x0001824428002844ULL,
-    0x0102048850005088ULL, 0x02040810A000A010ULL, 0x0408102040004020ULL,
-    0x1008040200020408ULL, 0x2010080500050810ULL, 0x4020110A000A1120ULL,
-    0x8041221400142241ULL, 0x0182442800284482ULL, 0x0204885000508804ULL,
-    0x040810A000A01008ULL, 0x0810204000402010ULL, 0x0804020002040810ULL,
-    0x1008050005081020ULL, 0x20110A000A112040ULL, 0x4122140014224180ULL,
-    0x8244280028448201ULL, 0x0488500050880402ULL, 0x0810A000A0100804ULL,
-    0x1020400040201008ULL, 0x0402000204081020ULL, 0x0805000508102040ULL,
-    0x110A000A11204080ULL, 0x2214001422418000ULL, 0x4428002844820100ULL,
-    0x8850005088040201ULL, 0x10A000A010080402ULL, 0x2040004020100804ULL,
-    0x0200020408102040ULL, 0x0500050810204080ULL, 0x0A000A1120408000ULL,
-    0x1400142241800000ULL, 0x2800284482010000ULL, 0x5000508804020100ULL,
-    0xA000A01008040201ULL, 0x4000402010080402ULL, 0x0002040810204080ULL,
-    0x0005081020408000ULL, 0x000A112040800000ULL, 0x0014224180000000ULL,
-    0x0028448201000000ULL, 0x0050880402010000ULL, 0x00A0100804020100ULL,
-    0x0040201008040201ULL};
-
 static const std::pair<int, int> kRookDirections[] = {
     {1, 0}, {-1, 0}, {0, 1}, {0, -1}};
 
@@ -397,21 +348,6 @@ BitBoard MagicBitBoards::rook_attacks_table_[102400];
 BitBoard MagicBitBoards::bishop_attacks_table_[5248];
 
 MagicBitBoards::MagicBitBoards() {
-  // Initialize masks for all board squares.
-  for (unsigned square = 0; square < 64; square++) {
-    const BoardSquare b_sq(square);
-
-    // Calculate relevant occupancy masks by subtracting the board edges from
-    // the total attacks bitboards.
-    rook_magic_params_[square].mask_ =
-        (kRookAttacks[square] - BoardSquare(b_sq.row(), 0) -
-         BoardSquare(b_sq.row(), 7) - BoardSquare(0, b_sq.col()) -
-         BoardSquare(7, b_sq.col()))
-            .as_int();
-    bishop_magic_params_[square].mask_ =
-        (kBishopAttacks[square] - BitBoard(0xFF818181818181FFULL)).as_int();
-  }
-
   // Build attacks tables.
   BuildAttacksTable(kRookMagicNumbers, rook_magic_params_, rook_attacks_table_,
                     kRookDirections);
@@ -429,6 +365,30 @@ void MagicBitBoards::BuildAttacksTable(const BitBoard* magic_numbers,
   // Initialize for all board squares.
   for (unsigned square = 0; square < 64; square++) {
     const BoardSquare b_sq(square);
+
+    // Calculate relevant occupancy masks by subtracting the board edges from
+    // the total attacks bitboards.
+    BitBoard mask = {0};
+
+    for (int j = 0; j < 4; j++) {
+      auto direction = directions[j];
+      auto dst_row = b_sq.row();
+      auto dst_col = b_sq.col();
+      while (true) {
+        dst_row += direction.first;
+        dst_col += direction.second;
+        // If the next square in this direction is invalid, the current square
+        // is at the board's edge and should not be added.
+        if (!BoardSquare::IsValid(dst_row + direction.first,
+                                  dst_col + direction.second))
+          break;
+        const BoardSquare destination(dst_row, dst_col);
+        mask.set(destination);
+      }
+    }
+
+    // Set mask.
+    magic_params[square].mask_ = mask.as_int();
 
     // Cache relevant occupancy board squares.
     std::vector<BoardSquare> occupancy_squares;