… src/chess/board.cc
… src/chess/board_test.cc
… src/chess/position_test.cc
… src/engine.cc
… src/main.cc
… src/mcts/search.cc
… src/neural/encoder.cc
… src/neural/encoder_test.cc
… src/syzygy/syzygy.cc
… src/syzygy/syzygy_test.cc