This PR adds magic bitboards (introduced in #628) as a layer on top of #638.

Unconditional speedup for random backend up to 10% (tested with LTO builds on Linux Fedora 29 64-bit, `gcc 8.2.1`).

The measurements details (command line and considered positions) are given in #638.

For 1 search thread:

position  |  master `6a639b6` |  #638  + this PR  | speedup
------ | ----- | ----------- | -----------
start | 207600 ± 447 | 216966 ± 414 | 4.5%
opening | 161055 ± 188 | 171825 ± 221 | 6.7%
endgame | 175728 ± 808 | 193532 ± 354 | 10%

For 2 search threads:

position  |  master `6a639b6` |  #638  + this PR  | speedup
------ | ----- | ----------- | -----------
start | 268477 ± 590 | 272720 ± 717 | 1.6%
opening | 221515 ± 461 | 228676 ± 491 | 3.2%
endgame | 235705 ± 632 | 245195 ± 655 | 4.0%