@ddobbelaere 2019-01-06
On my machine `chessboard_test` (built with `b_lto=true`) with same uncommented tests as [#638 (comment)](https://github.com/LeelaChessZero/lc0/pull/638#issuecomment-451684440):

current master (includes #638):

```
[ RUN      ] ChessBoard.MoveGenStartingPos
[       OK ] ChessBoard.MoveGenStartingPos (8084 ms)
[ RUN      ] ChessBoard.MoveGenKiwipete
[       OK ] ChessBoard.MoveGenKiwipete (14440 ms)
[ RUN      ] ChessBoard.MoveGenPosition3
[       OK ] ChessBoard.MoveGenPosition3 (1077 ms)
[ RUN      ] ChessBoard.MoveGenPosition4
[       OK ] ChessBoard.MoveGenPosition4 (1300 ms)
[ RUN      ] ChessBoard.MoveGenPosition5
[       OK ] ChessBoard.MoveGenPosition5 (6128 ms)
[ RUN      ] ChessBoard.MoveGenPosition6
[       OK ] ChessBoard.MoveGenPosition6 (262 ms)
...
[----------] 15 tests from ChessBoard (31292 ms total)
```
this PR (without PEXT):
```
[ RUN      ] ChessBoard.MoveGenStartingPos
[       OK ] ChessBoard.MoveGenStartingPos (7512 ms)
[ RUN      ] ChessBoard.MoveGenKiwipete
[       OK ] ChessBoard.MoveGenKiwipete (10888 ms)
[ RUN      ] ChessBoard.MoveGenPosition3
[       OK ] ChessBoard.MoveGenPosition3 (838 ms)
[ RUN      ] ChessBoard.MoveGenPosition4
[       OK ] ChessBoard.MoveGenPosition4 (972 ms)
[ RUN      ] ChessBoard.MoveGenPosition5
[       OK ] ChessBoard.MoveGenPosition5 (5320 ms)
[ RUN      ] ChessBoard.MoveGenPosition6
[       OK ] ChessBoard.MoveGenPosition6 (218 ms)
...
[----------] 15 tests from ChessBoard (25749 ms total)
```
This PR (with PEXT, built with `pext=true`):
```
[ RUN      ] ChessBoard.MoveGenStartingPos
[       OK ] ChessBoard.MoveGenStartingPos (7371 ms)
[ RUN      ] ChessBoard.MoveGenKiwipete
[       OK ] ChessBoard.MoveGenKiwipete (10748 ms)
[ RUN      ] ChessBoard.MoveGenPosition3
[       OK ] ChessBoard.MoveGenPosition3 (821 ms)
[ RUN      ] ChessBoard.MoveGenPosition4
[       OK ] ChessBoard.MoveGenPosition4 (960 ms)
[ RUN      ] ChessBoard.MoveGenPosition5
[       OK ] ChessBoard.MoveGenPosition5 (5225 ms)
[ RUN      ] ChessBoard.MoveGenPosition6
[       OK ] ChessBoard.MoveGenPosition6 (215 ms)
...
[----------] 15 tests from ChessBoard (25340 ms total)
```