@ddobbelaere 2019-01-06
This PR is ready for review.