@ddobbelaere 2019-01-06
I thought the current way of using a constructor was quite neat (I was a
bit in Java mindset where static initialization is very easy and common,
even without this constructor trick), but this fails miserably if the dummy
instance gets optimized away.

I like your last suggestion to move everything into anonymous namespace of
board.cc (dropping the class on top) and initialize explicitly. This has
the added advantage that we have full control where to do the
initialization (it will probably take only few milliseconds, but still...).
I'll implement the changes for this last approach in the coming days.

Op zo 6 jan. 2019 13:03 schreef Alexander Lyashuk <notifications@github.com:

> *@mooskagh* commented on this pull request.
> ------------------------------
>
> In src/chess/bitboard.h
> <https://github.com/LeelaChessZero/lc0/pull/640#discussion_r245505446>:
>
> > @@ -229,6 +238,86 @@ class BitBoard {
>    std::uint64_t board_ = 0;
>  };
>
> +// Holds magic bitboard routines and performs initialization.
> +// We use so-called "fancy" magic bitboards.
> +class MagicBitBoards {
> + public:
> +  // Constructor is called once in bitboard.cc to initialize structures, should
> +  // not be called again!
> +  MagicBitBoards();
>
> Does using a singleton instead make it much slower?
> E.g.
>
> class MagicBitBoards {
>  public:
>    static const MagicBitBoards& Get() {
>        static MagicBitBoards boards_;
>        return boards;
>    }
>  private:
>   MagicBitBoards() { ... }
> };
>
> (and removing static from everywhere)
>
> It would be neater (I guess), but I can see that it could be slower
> indeed..
>
> On other hand, unused static variable can be optimized out, so the
> constructor will never run.
> (we had this problem with backend registration btw, which are also unused
> static variables,
> after some massaging linker now doesn't optimize them, but it didn't
> before so it's fragile)
> ------------------------------
>
> In src/chess/bitboard.h
> <https://github.com/LeelaChessZero/lc0/pull/640#discussion_r245505592>:
>
> > @@ -229,6 +238,86 @@ class BitBoard {
>    std::uint64_t board_ = 0;
>  };
>
> +// Holds magic bitboard routines and performs initialization.
> +// We use so-called "fancy" magic bitboards.
> +class MagicBitBoards {
> + public:
> +  // Constructor is called once in bitboard.cc to initialize structures, should
> +  // not be called again!
> +  MagicBitBoards();
>
> If it's not slower, then doing MagicBitBoards::Get() in main() so that
> time is not spent during actual game is fine.
> ------------------------------
>
> In src/chess/bitboard.h
> <https://github.com/LeelaChessZero/lc0/pull/640#discussion_r245505763>:
>
> > @@ -229,6 +238,86 @@ class BitBoard {
>    std::uint64_t board_ = 0;
>  };
>
> +// Holds magic bitboard routines and performs initialization.
> +// We use so-called "fancy" magic bitboards.
> +class MagicBitBoards {
> + public:
> +  // Constructor is called once in bitboard.cc to initialize structures, should
> +  // not be called again!
> +  MagicBitBoards();
>
> If it is slower, then still calling some InitializeMagicBitboards() from
> main() seems to be more robust approach.
>
> In that case, maybe it's better to move everything into anonymous
> namespace inside board.cc (not bitboard.cc, for me magicbitboards are
> more part of movegen than bitboard itself, it feels similar only because of
> name), and only expose one function InitializeMagicBitboards() in .h
> file. Also in this case the class wrapper (which only serves purpose of
> namespace) is not necessary I think.
> ------------------------------
>
> In src/chess/bitboard.h
> <https://github.com/LeelaChessZero/lc0/pull/640#discussion_r245505867>:
>
> > @@ -229,6 +238,86 @@ class BitBoard {
>    std::uint64_t board_ = 0;
>  };
>
> +// Holds magic bitboard routines and performs initialization.
> +// We use so-called "fancy" magic bitboards.
> +class MagicBitBoards {
> + public:
> +  // Constructor is called once in bitboard.cc to initialize structures, should
> +  // not be called again!
> +  MagicBitBoards();
>
> Actually what I wrote in the last paragraph I like more then singleton
> even..
> But I wonder what's your opinion.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/LeelaChessZero/lc0/pull/640#pullrequestreview-189626555>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/Ak9B930z185lePebQUi08LZVcJIwVt55ks5vAeX-gaJpZM4Zl2is>
> .
>